requirements/dev.txt
# Наследуем основные зависимости
-r base.txt

#############################################
# 1. Инструменты для удобной разработки
#############################################

# Улучшенный вывод тестов
pytest-sugar==0.9.7          # Красивые прогресс-бары
pytest-instafail==0.5.0      # Показ упавших тестов сразу
pytest-pretty==1.2.0         # Цветной вывод

# Отладка
ipython==8.14.0              # Улучшенная консоль
debugpy==1.8.0               # Дебаггер для VSCode
pudb==2022.1.3               # TUI-дебаггер

#############################################
# 2. Генерация тестовых данных
#############################################
factory-boy==3.3.0           # Фабрики объектов
fake-barcode==0.9.0          # Генерация тестовых данных
mimesis==12.1.0              # Реалистичные фейковые данные

#############################################
# 3. Линтеры и качество кода
#############################################
flake8==6.1.0                # Проверка стиля
flake8-bugbear==23.7.10      # Поиск потенциальных багов
mypy==1.5.1                  # Статическая типизация
black==23.7.0                # Форматирование кода
isort==5.12.0                # Сортировка импортов

#############################################
# 4. Дополнительные инструменты Playwright
#############################################
playwright-codegen==1.39.0   # Генератор кода из действий
playwright-test==0.2.2       # Доп. утилиты для тестов

#############################################
# 5. Документирование
#############################################
pytest-docstyle==1.1.0       # Проверка docstrings
sphinx==7.2.5                # Генерация документации